# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "summary": "Get all users",
        "description": "Returns a paginated list of all users",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of type User",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PageModel"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getUser",
        "summary": "Get user by ID",
        "description": "Returns a user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/me": {
      "get": {
        "operationId": "SessionUserController_getSessionUser",
        "summary": "Get current user",
        "description": "Returns the current user. Requires a valid JWT token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Unauthorized",
                  "statusCode": 401
                }
              }
            }
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/logout": {
      "get": {
        "operationId": "SessionUserController_logout",
        "summary": "Logout",
        "description": "Clears the JWT cookie, and invalidates the JWT token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The user has been logged out successfully."
          },
          "401": {
            "description": "If the user is not authenticated, it cannot log out."
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/settings": {
      "get": {
        "operationId": "SessionUserController_getSessionUserSettings",
        "summary": "Get current user settings",
        "description": "Returns the current user settings. Requires a valid JWT token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SessionUserController_updateSessionUserSettings",
        "summary": "Update current user settings",
        "description": "Updates the current user settings. Requires a valid JWT token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettings"
                }
              }
            }
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/integrations": {
      "get": {
        "operationId": "SessionUserController_getSessionUserIntegrations",
        "summary": "Get current user integrations",
        "description": "Will return the currently integrated cloud git systems",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/integrations/{system}": {
      "get": {
        "operationId": "SessionUserController_getSessionUserIntegration",
        "summary": "Get current user integration",
        "description": "Will return the currently integrated cloud git system with related repositories",
        "parameters": [
          {
            "name": "system",
            "required": true,
            "in": "query",
            "description": "The cloud git system to get the available repositories for",
            "schema": {
              "enum": [
                "github",
                "gitlab",
                "bitbucket"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/integrations/all/available": {
      "get": {
        "operationId": "SessionUserController_getSessionUserAllAvailableRepositories",
        "summary": "Get available repositories",
        "description": "Will return the available repositories for all cloud git systems",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/integrations/{system}/available": {
      "get": {
        "operationId": "SessionUserController_getSessionUserAvailableRepositories",
        "summary": "Get available repositories",
        "description": "Will return the available repositories for the given cloud git system",
        "parameters": [
          {
            "name": "system",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/installations": {
      "get": {
        "operationId": "SessionUserController_getSessionUserInstallations",
        "summary": "Get current user installations",
        "description": "Will return the currently installed GitHub Apps",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/install/{system}/{repoFullName}": {
      "post": {
        "operationId": "SessionUserController_installWebhookOnRepo",
        "summary": "Install webhook on repo",
        "description": "Will install a webhook on the given repository",
        "parameters": [
          {
            "name": "system",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repoFullName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SessionUserController_uninstallWebhookOnRepo",
        "summary": "Uninstall webhook on repo",
        "description": "Will uninstall a webhook on the given repository",
        "parameters": [
          {
            "name": "system",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repoFullName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/me/github/post-install": {
      "get": {
        "operationId": "SessionUserController_postInstallGithubApp",
        "summary": "Post install",
        "description": "Will post install the GitHub App",
        "parameters": [
          {
            "name": "installation_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "setup_action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Session User"
        ],
        "security": [
          {
            "cookie": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "achievibit-api",
    "description": "![swagger-mode](https://img.shields.io/badge/mode-devcontainer-FF5BF8)\n\nSince this swagger shares the same domain as the app, you can use the same cookie for authentication.\n\nFor the WebSocket API, please visit [achievibit-ws](/api/docs-async)",
    "version": "1.0",
    "contact": {
      "name": "thatkookooguy",
      "url": "https://github.com/thatkookooguy",
      "email": "thatkookooguy@kibibit.io"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT token for authenticated user routes"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "user_token",
        "description": "JWT token for authenticated user routes"
      },
      "gitlab-webhook": {
        "type": "apiKey",
        "name": "x-gitlab-token",
        "in": "header",
        "description": "Secret token for GitLab webhooks validation (usually sent by GitLab)"
      },
      "github-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature-256",
        "in": "header",
        "description": "Secret token for GitHub webhooks validation (usually sent by GitHub)"
      },
      "bitbucket-webhook": {
        "type": "apiKey",
        "name": "x-hub-signature",
        "in": "header",
        "description": "Secret token for Bitbucket webhooks validation (usually sent by Bitbucket)"
      }
    },
    "schemas": {
...
      "PageMetaModel": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "take": {
            "type": "number"
          },
          "itemCount": {
            "type": "number"
          },
          "pageCount": {
            "type": "number"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "take",
          "itemCount",
          "pageCount",
          "hasPreviousPage",
          "hasNextPage"
        ]
      },
      "PageModel": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaModel"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "Integration": {
        "type": "object",
        "properties": {
          "system": {
            "type": "string",
            "enum": [
              "github",
              "gitlab",
              "bitbucket"
            ]
          },
          "systemEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "systemUsername": {
            "type": "string"
          },
          "systemAvatar": {
            "type": "string"
          },
          "organizations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "system",
          "systemEmails",
          "systemUsername",
          "systemAvatar",
          "organizations"
        ]
      },
      "UserSettings": {
        "type": "object",
        "properties": {
          "timezone": {
            "type": "string"
          },
          "dateFormat": {
            "type": "string"
          },
          "avatarSystemOrigin": {
            "type": "string",
            "enum": [
              "github",
              "gitlab",
              "bitbucket"
            ]
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ]
          }
        },
        "required": [
          "timezone",
          "dateFormat",
          "avatarSystemOrigin",
          "theme"
        ]
      },
      "Achievement": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "pullRequestUrl": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "treasure": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "pullRequestUrl",
          "avatar",
          "name",
          "description",
          "treasure"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "user",
                "guest"
              ]
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "view:pull-requests"
              ]
            }
          },
          "lives": {
            "type": "number"
          },
          "integrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Integration"
            }
          },
          "settings": {
            "$ref": "#/components/schemas/UserSettings"
          },
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Organization"
            }
          },
          "achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Achievement"
            }
          },
          "registered": {
            "type": "boolean"
          }
        },
        "required": [
          "createdAt",
          "username",
          "avatar",
          "isOnboarded",
          "roles",
          "permissions",
          "lives",
          "integrations",
          "settings",
          "organizations",
          "achievements",
          "registered"
        ]
      },
      "Organization": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "system": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/User"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "required": [
          "createdAt",
          "name",
          "system",
          "owner",
          "members"
        ]
      },
...
    }
  }
}